@@ -1,17 +1,16 @@
-﻿using Entities.ExtendedModels;
-using Entities.Models;
+﻿using Entities.Models;
 using System;
 using System.Collections.Generic;
 
 namespace Contracts
 {
-    public interface IOwnerRepository:IRepositoryBase<Owner>
+    public interface IOwnerRepository : IRepositoryBase<Owner>
     {
         IEnumerable<Owner> GetAllOwners();
         Owner GetOwnerById(Guid ownerId);
-        OwnerExtended GetOwnerWithDetails(Guid ownerId);
+        Owner GetOwnerWithDetails(Guid ownerId);
         void CreateOwner(Owner owner);
-        void UpdateOwner(Owner dbOwner, Owner owner);
+        void UpdateOwner(Owner owner);
         void DeleteOwner(Owner owner);
     }
 }
