@@ -1,5 +1,6 @@
 ﻿using System;
 using System.Collections.Generic;
+using System.Linq;
 using System.Linq.Expressions;
 using System.Text;
 
@@ -7,11 +8,10 @@ namespace Contracts
 {
     public interface IRepositoryBase<T>
     {
-        IEnumerable<T> FindAll();
-        IEnumerable<T> FindByCondition(Expression<Func<T, bool>> expression);
-        void Create(T entity);
-        void Update(T entity);
+        IQueryable<T> FindAll(); 
+        IQueryable<T> FindByCondition(Expression<Func<T, bool>> expression); 
+        void Create(T entity); 
+        void Update(T entity); 
         void Delete(T entity);
-        void Save();
     }
 }
