@@ -1,26 +1,25 @@
-﻿using Entities.Enumerations;
-using System;
-using System.Collections.Generic;
+﻿using System;
 using System.ComponentModel.DataAnnotations;
 using System.ComponentModel.DataAnnotations.Schema;
-using System.Text;
 
 namespace Entities.Models
 {
-    [Table("account")]
-    public class Account : IEntity
+    [Table("account")] 
+    public class Account 
     {
-        [Key]
         [Column("AccountId")]
         public Guid Id { get; set; }
 
-        [Required(ErrorMessage = "Date created is required")]
+        [Required(ErrorMessage = "Date created is required")] 
         public DateTime DateCreated { get; set; }
 
-        [Required(ErrorMessage = "Account type is required")]
+        [Required(ErrorMessage = "Account type is required")] 
         public string AccountType { get; set; }
 
-        [Required(ErrorMessage = "Owner Id is required")]
+        [Required(ErrorMessage = "Owner Id is required")] 
+
+        [ForeignKey(nameof(Owner))]
         public Guid OwnerId { get; set; }
+        public Owner Owner { get; set; }
     }
 }
