@@ -2,26 +2,26 @@
 using System.Collections.Generic;
 using System.ComponentModel.DataAnnotations;
 using System.ComponentModel.DataAnnotations.Schema;
-using System.Text;
 
 namespace Entities.Models
 {
-    [Table("owner")]
-    public class Owner : IEntity
-    {
-        [Key]
+    [Table("owner")] 
+    public class Owner 
+    { 
         [Column("OwnerId")]
         public Guid Id { get; set; }
 
-        [Required(ErrorMessage = "Name is required")]
-        [StringLength(60, ErrorMessage = "Name can't be longer than 60 characters")]
+        [Required(ErrorMessage = "Name is required")] 
+        [StringLength(60, ErrorMessage = "Name can't be longer than 60 characters")] 
         public string Name { get; set; }
 
-        [Required(ErrorMessage = "Date of birth is required")]
+        [Required(ErrorMessage = "Date of birth is required")] 
         public DateTime DateOfBirth { get; set; }
 
-        [Required(ErrorMessage = "Address is required")]
-        [StringLength(100, ErrorMessage = "Address can not be loner then 100 characters")]
+        [Required(ErrorMessage = "Address is required")] 
+        [StringLength(100, ErrorMessage = "Address cannot be loner then 100 characters")] 
         public string Address { get; set; }
+
+        public ICollection<Account> Accounts { get; set; }
     }
 }
