@@ -3,11 +3,11 @@ using Microsoft.EntityFrameworkCore;
 
 namespace Entities
 {
-    public class RepositoryContext: DbContext
+    public class RepositoryContext : DbContext
     {
-        public RepositoryContext(DbContextOptions options)
-            :base(options)
-        {
+        public RepositoryContext(DbContextOptions options) 
+            : base(options) 
+        { 
         }
 
         public DbSet<Owner> Owners { get; set; }
