@@ -3,19 +3,20 @@ using Entities;
 using Entities.Models;
 using System;
 using System.Collections.Generic;
+using System.Linq;
 
 namespace Repository
 {
-    public class AccountRepository : RepositoryBase<Account>, IAccountRepository
-    {
-        public AccountRepository(RepositoryContext repositoryContext)
-            : base(repositoryContext)
-        {
+    public class AccountRepository : RepositoryBase<Account>, IAccountRepository 
+    { 
+        public AccountRepository(RepositoryContext repositoryContext) 
+            : base(repositoryContext) 
+        { 
         }
 
         public IEnumerable<Account> AccountsByOwner(Guid ownerId)
-        {
-            return FindByCondition(a => a.OwnerId.Equals(ownerId));
+        { 
+            return FindByCondition(a => a.OwnerId.Equals(ownerId)).ToList(); 
         }
     }
 }
