@@ -1,61 +1,53 @@
-﻿using System;
+﻿using Contracts;
+using Entities;
+using Entities.Models;
+using Microsoft.EntityFrameworkCore;
+using System;
 using System.Collections.Generic;
 using System.Linq;
-using Contracts;
-using Entities;
-using Entities.ExtendedModels;
-using Entities.Extensions;
-using Entities.Models;
 
 namespace Repository
 {
-    public class OwnerRepository : RepositoryBase<Owner>, IOwnerRepository
-    {
-        public OwnerRepository(RepositoryContext repositoryContext)
-            : base(repositoryContext)
-        {
+    public class OwnerRepository : RepositoryBase<Owner>, IOwnerRepository 
+    { 
+        public OwnerRepository(RepositoryContext repositoryContext) 
+            : base(repositoryContext) 
+        { 
         }
 
-        public IEnumerable<Owner> GetAllOwners()
-        {
+        public IEnumerable<Owner> GetAllOwners() 
+        { 
             return FindAll()
-                .OrderBy(ow => ow.Name);
+                .OrderBy(ow => ow.Name)
+                .ToList(); 
         }
 
         public Owner GetOwnerById(Guid ownerId)
         {
             return FindByCondition(owner => owner.Id.Equals(ownerId))
-                .DefaultIfEmpty(new Owner())
                 .FirstOrDefault();
         }
 
-        public OwnerExtended GetOwnerWithDetails(Guid ownerId)
+        public Owner GetOwnerWithDetails(Guid ownerId)
         {
-            return new OwnerExtended(GetOwnerById(ownerId))
-            {
-                Accounts = RepositoryContext.Accounts
-                    .Where(a => a.OwnerId == ownerId)
-            };
+            return FindByCondition(owner => owner.Id.Equals(ownerId))
+                .Include(ac => ac.Accounts)
+                .FirstOrDefault();
         }
 
         public void CreateOwner(Owner owner)
         {
-            owner.Id = Guid.NewGuid();
             Create(owner);
-            Save();
         }
 
-        public void UpdateOwner(Owner dbOwner, Owner owner)
+        public void UpdateOwner(Owner owner)
         {
-            dbOwner.Map(owner);
-            Update(dbOwner);
-            Save();
+            Update(owner);
         }
 
         public void DeleteOwner(Owner owner)
         {
             Delete(owner);
-            Save();
         }
     }
 }
